moving lineContainerComponent
