+++ guiEditor/src/components/guiList/guiListComponent.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { GlobalState } from "../../globalState";
-import { LineContainerComponent } from "../../sharedComponents/lineContainerComponent";
+import { LineContainerComponent } from "../../sharedUiComponents/lines/lineContainerComponent";
 import { DraggableLineComponent } from "../../sharedUiComponents/lines/draggableLineComponent";
 import { Observer } from "babylonjs/Misc/observable";
 import { Nullable } from "babylonjs/types";
